ci: run new workflow for github actions
I chaged the flows of pull requests and tagging step. Before pre-release was working when we merge a
branch to the dev. Now if we want to get a pre-release we should trigger it by the chosing the
branch on Actions. Before release step was working when we get a new tag, now after pre-release we
will have a release-pr from pre-release process. and when we merge it to the master then release
process will be run on the actions